# voice-studio-openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Aether Voice Gateway",
    "version": "1.0.0"
  },
  "paths": {
...
    "/v1/asr/transcribe": {
      "post": {
        "tags": [
          "asr"
        ],
        "summary": "Transcribe",
        "operationId": "transcribe_v1_asr_transcribe_post",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ASRTranscribeResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/asr/stream/start": {
      "post": {
        "tags": [
          "asr"
        ],
        "summary": "Start Stream",
        "operationId": "start_stream_v1_asr_stream_start_post",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ASRStreamStartRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ASRStreamStartResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ASRSegment": {
        "properties": {
          "segment_id": {
            "type": "string",
            "title": "Segment Id"
          },
          "start_ms": {
            "type": "integer",
            "title": "Start Ms"
          },
          "end_ms": {
            "type": "integer",
            "title": "End Ms"
          },
          "text": {
            "type": "string",
            "title": "Text"
          },
          "confidence": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Confidence"
          }
        },
        "type": "object",
        "required": [
          "segment_id",
          "start_ms",
          "end_ms",
          "text"
        ],
        "title": "ASRSegment"
      },
      "ASRStreamStartRequest": {
        "properties": {
          "model": {
            "type": "string",
            "title": "Model",
            "default": "auto"
          },
          "language": {
            "type": "string",
            "title": "Language",
            "default": "auto"
          },
          "sample_rate": {
            "type": "integer",
            "title": "Sample Rate",
            "default": 16000
          },
          "encoding": {
            "type": "string",
            "title": "Encoding",
            "default": "pcm_s16le"
          },
          "channels": {
            "type": "integer",
            "title": "Channels",
            "default": 1
          },
          "triage_enabled": {
            "type": "boolean",
            "title": "Triage Enabled",
            "default": false
          },
          "metadata": {
            "$ref": "#/components/schemas/RequestMetadata"
          }
        },
        "type": "object",
        "title": "ASRStreamStartRequest"
      },
      "ASRStreamStartResponse": {
        "properties": {
          "session_id": {
            "type": "string",
            "title": "Session Id"
          },
          "ws_url": {
            "type": "string",
            "title": "Ws Url"
          },
          "expires_in_seconds": {
            "type": "integer",
            "title": "Expires In Seconds",
            "default": 3600
          },
          "model_requested": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Model Requested"
          },
          "model_used": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Model Used"
          },
          "fallback_used": {
            "type": "boolean",
            "title": "Fallback Used",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "session_id",
          "ws_url"
        ],
        "title": "ASRStreamStartResponse"
      },
      "ASRTranscribeResponse": {
        "properties": {
          "request_id": {
            "type": "string",
            "title": "Request Id"
          },
          "session_id": {
            "type": "string",
            "title": "Session Id"
          },
          "task": {
            "type": "string",
            "title": "Task"
          },
          "model_requested": {
            "type": "string",
            "title": "Model Requested"
          },
          "model_used": {
            "type": "string",
            "title": "Model Used"
          },
          "language_detected": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Language Detected"
          },
          "duration_ms": {
            "type": "integer",
            "title": "Duration Ms",
            "default": 0
          },
          "text": {
            "type": "string",
            "title": "Text"
          },
          "segments": {
            "items": {
              "$ref": "#/components/schemas/ASRSegment"
            },
            "type": "array",
            "title": "Segments"
          },
          "timings": {
            "$ref": "#/components/schemas/TimingBreakdown"
          },
          "artifacts": {
            "additionalProperties": true,
            "type": "object",
            "title": "Artifacts"
          }
        },
        "type": "object",
        "required": [
          "request_id",
          "session_id",
          "task",
          "model_requested",
          "model_used",
          "text",
          "timings"
        ],
        "title": "ASRTranscribeResponse"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "RequestMetadata": {
        "properties": {
          "source": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source"
          },
          "tenant_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tenant Id"
          },
          "trace_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Trace Id"
          },
          "extra": {
            "additionalProperties": true,
            "type": "object",
            "title": "Extra"
          }
        },
        "type": "object",
        "title": "RequestMetadata"
      },
...
      "TimingBreakdown": {
        "properties": {
          "queue_ms": {
            "type": "integer",
            "title": "Queue Ms",
            "default": 0
          },
          "preprocess_ms": {
            "type": "integer",
            "title": "Preprocess Ms",
            "default": 0
          },
          "inference_ms": {
            "type": "integer",
            "title": "Inference Ms",
            "default": 0
          },
          "postprocess_ms": {
            "type": "integer",
            "title": "Postprocess Ms",
            "default": 0
          },
          "total_ms": {
            "type": "integer",
            "title": "Total Ms",
            "default": 0
          },
          "encode_ms": {
            "type": "integer",
            "title": "Encode Ms",
            "default": 0
          }
        },
        "type": "object",
        "title": "TimingBreakdown"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}